backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
boto3==1.35.0
python-dotenv==1.0.0
httpx==0.27.0
numpy==1.26.4
pdfplumber==0.11.4
redis==5.0.8
python-multipart==0.0.9
aiofiles==24.1.0
reportlab==4.2.2
nova-act
gtts
aws-sdk-bedrock-runtime
smithy-aws-core
smithy-core
